Refactored the datepicker interactor into it's own file
