+++ bigtest/interactors/create-page.js
@@ -4,37 +4,11 @@ import {
   interactor,
   isPresent,
   property,
-  selectable,
   text,
   value,
-  collection,
-  focusable,
-  blurrable,
-  action,
   scoped,
-  isVisible
 } from '@bigtest/interactor';
-
-@interactor class DatePicker {
-  blurInput = blurrable('input');
-  focusInput = clickable('input');
-  fillInput = fillable('input');
-  value = value('input');
-  isDayPickerVisible = isVisible('.DayPicker');
-
-  dayCells = collection('.DayPicker-Day', {
-    clickDay: clickable(),
-    dayText: text()
-  });
-
-  async selectDay(day) {
-    await this.focusInput()
-      .when(() => this.isDayPickerVisible)
-    await this.dayCells()
-      .find(cell => cell.dayText === `${day}`)
-      .clickDay();
-  }
-}
+import DatePicker from './date-picker';
 
 @interactor class CreatePage {
   hasHeading = isPresent('h6');
@@ -47,9 +21,6 @@ import {
   nameIsReadOnly = property('[data-test-owner-name]', 'readOnly');
 
   changeOwnerName = fillable('[data-test-owner-name]');
-
-  // changeStartDate = fillable('[data-test-start-date]');
-  // changeEndDate = fillable('[data-test-end-date]');
 
   changeStartDate(day) {
     return this.startDate.selectDay(day);
